+++ tests/pytorch/mock_sparse2/test_sddmm.py
@@ -4,13 +4,8 @@ import backend as F
 import pytest
 import torch
 from dgl.mock_sparse2 import sddmm
-from utils import (
-    clone_detach_and_grad,
-    rand_coo,
-    rand_csc,
-    rand_csr,
-    sparse_matrix_to_dense,
-)
+
+from .utils import clone_detach_and_grad, rand_coo, rand_csc, rand_csr
 
 # TODO(#4818): Skipping tests on win.
 if not sys.platform.startswith("linux"):
